Bug fix for `automatic_serialization` in Zend
This bug fix addresses an issue reported by @kVasiliev involving
`Zend_Cache_Core` with `automatic_serialization = true` which no unit
tests existed for Backend testing with this setting. Once the setting
was enabled it was discovered the
`Rediska_Command_GetFromHash::parseResponse` unserializes all responses.
To control this behavior a property and public method was added to the
`Rediska_Serializer`.
Property: `Rediska_Serializer::_serializationEnabled`
Method: `Rediska_Serializer::toggleSerialization` to allow runtime
enable and disable of this behavior. Calls to this method occur in the
`Rediska_Zend_Cache_Backend_Redis::load` allowing for the return of an
unserialized string. Because all things being passed in and out of
`Zend_Cache` should by nature be serialized this should aleviate any
further issue. While I am not comfortable with this change effecting the
`Rediska_Serializer` I have not at this time found a means by which to
address this issue and not at the same time inject potential BC
breakage.